requirements.txt
Flask
jsonify 
requests 
python-dotenv